requirements.txt
pygame
openai
python-dotenv 